change structure and add testing
