+++ .github/workflows/tests.yml
@@ -1,0 +1,60 @@
+name: Python Unit Tests
+
+
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+    - main
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      -
+        name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        # This is the version of the action for setting up Python, not the Python version.
+        uses: actions/setup-python@v5
+        with:
+          # Semantic version range syntax or exact version of a Python version
+          python-version: '3.10.12'
+          # Optional - x64 or x86 architecture, defaults to x64
+          architecture: 'x64'
+      
+      - name: Display Python version
+        run: python -c "import sys; print(sys.version)"
+      
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pytest httpx
+          pip install -r requirements.txt
+      
+      - name: install apt dependecies
+        run:  |
+          sudo apt-get update
+          sudo apt-get install -y build-essential autoconf automake libtool pkg-config git wget cmake
+          rm -rf /var/lib/apt/lists/*
+      
+      - name: install espeak-ng
+        run:  |
+          git clone -b dev-ca https://github.com/projecte-aina/espeak-ng
+          pip install --upgrade pip 
+          cd espeak-ng && ./autogen.sh && ./configure --prefix=/usr && make && make install
+
+      - name: install lingua-franca
+        run: pip install git+https://github.com/MycroftAI/lingua-franca.git@5bfd75fe5996fd364102a0eec3f714c9ddc9275c
+
+      - name: install model
+        run:  |
+          wget -q http://share.laklak.eu/model_vits_ca/best_model_8khz.pth -P ./models/vits_ca/
+          mv ./models/vits_ca/best_model_8khz.pth ./models/vits_ca/best_model.pth
+
+      - name: Run tests
+        run: pytest
+
+
+
