+++ docker-compose-test.yml
@@ -1,0 +1,14 @@
+version: '3.9'
+services:
+  server:
+    build:
+      context: .
+      dockerfile: Dockerfile.test
+    container_name: fastapi-test
+    restart: always
+    volumes:
+      - .:/app
+
+    ports:
+      - '8080:8000'
+    shm_size: ${SHM_SIZE-2gb}
