+++ server/exceptions.py
@@ -1,0 +1,8 @@
+class SpeakerException(Exception):
+    def __init__(self, speaker_id: str):
+        self.speaker_id = speaker_id
+
+
+class LanguageException(Exception):
+    def __init__(self, language: str):
+        self.language = language
