+++ server/helper/singleton.py
@@ -1,0 +1,16 @@
+from typing import Dict
+
+
+class Singleton(type):
+    _instances: Dict = {}
+
+    def __call__(cls, *args, **kwargs):
+        if kwargs:
+            cls._instances = {}
+
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(
+                *args, **kwargs
+            )
+
+        return cls._instances[cls]
